pyproject.toml
[project]
name = "open-trade"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "dhanhq>=2.0.0",
    "pandas-ta>=0.3.14b0",
    "yfinance>=0.2.48",
    "feedparser>=6.0.11",
    "pandas>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "python-telegram-bot>=21.0",
    "pytz>=2024.1",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sse-starlette>=1.8.0",
]

[tool.setuptools.packages.find]
include = ["agent*", "data*", "risk*", "api*"]

[project.scripts]
trade = "agent.main:entry"
api = "api.server:start"
